+++ resources/views/errors/401.blade.php
@@ -1,0 +1,5 @@
+@extends('errors.layout')
+
+@section('code', '401')
+@section('title', __('Please sign in'))
+@section('message', __('This page needs you to be signed in. Your session may have ended while the tab was open.'))