+++ resources/views/errors/403.blade.php
@@ -1,0 +1,5 @@
+@extends('errors.layout')
+
+@section('code', '403')
+@section('title', __('You do not have access to this'))
+@section('message', __('Your account is not permitted to open this page. If you need it, ask whoever manages your permissions.'))