+++ resources/views/errors/404.blade.php
@@ -1,0 +1,5 @@
+@extends('errors.layout')
+
+@section('code', '404')
+@section('title', __('Page not found'))
+@section('message', __('This page does not exist, or it was moved. Check the address, or go back to where you were.'))