+++ resources/views/errors/419.blade.php
@@ -1,0 +1,5 @@
+@extends('errors.layout')
+
+@section('code', '419')
+@section('title', __('This page expired'))
+@section('message', __('You had the form open long enough for the page to go stale. Open it again and your details will be there to re-enter.'))