+++ resources/views/errors/429.blade.php
@@ -1,0 +1,5 @@
+@extends('errors.layout')
+
+@section('code', '429')
+@section('title', __('Too many requests'))
+@section('message', __('You have sent a lot of requests in a short time. Wait a moment and try again.'))